pyproject.toml
[project]
name = "retrieval-augmented-transformer"
version = "1.0.0"
description = "Retrieval-Augmented Transformer for Question Answering"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "rouge-score>=0.1.2",
    "nltk>=3.8.0",
    "accelerate>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

